requirements.txt
absl-py==2.1.0
accelerate==1.10.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.1.2
annotated-types==0.7.0
anthropic==0.61.0
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.0
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.9.0
Brotli==1.1.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
executing==2.0.0
fastapi==0.116.1
fastjsonschema==2.18.1
ffmpy==0.6.1
filelock==3.18.0
filetype==1.2.0
flash==1.0.3
Flask==3.1.1
Flask-PyMongo==3.0.1
flatbuffers==24.3.25
fonttools==4.51.0
fpdf==1.7.2
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.7.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.40
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
gradio==5.41.0
gradio_client==1.11.0
greenlet==3.2.4
groovy==0.1.2
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
h5py==3.10.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.3
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.16.1
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.2
jiter==0.10.0
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.2
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
keras==3.2.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
langchain==0.3.27
langchain-anthropic==0.3.18
langchain-community==0.3.27
langchain-core==0.3.73
langchain-google-genai==2.0.10
langchain-tavily==0.2.11
langchain-text-splitters==0.3.9
langsmith==0.4.13
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
markdown2==2.5.4
MarkupSafe==2.1.3
marshmallow==3.26.1
matplotlib==3.8.4
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
namex==0.0.7
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.5
nltk==3.8.1
notebook==7.0.4
notebook_shim==0.2.3
numpy==2.3.2
openai==1.99.1
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
orjson==3.11.1
overrides==7.4.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.11.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==5.9.5
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==14.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.21
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.8.1b0
pydub==0.25.1
pygame==2.5.1
Pygments==2.16.1
pymongo==4.14.0
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.1.1
python-json-logger==2.0.7
python-multipart==0.0.20
pytz==2023.3
PyWavefront==1.3.3
pywin32==306
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
regex==2023.10.3
requests==2.32.4
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rpds-py==0.10.3
rsa==4.9.1
ruff==0.12.7
safehttpx==0.1.6
safetensors==0.6.1
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.2
shellingham==1.5.4
six==1.16.0
sklearn==0.0.post9
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.42
stack-data==0.6.3
starlette==0.47.2
streamlit==1.28.1
sympy==1.14.0
tavily-python==0.7.10
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.17.1
threadpoolctl==3.2.0
tiktoken==0.10.0
tinycss2==1.2.1
tokenizers==0.21.4
toml==0.10.2
tomlkit==0.13.3
toolz==0.12.0
torch==2.7.1
torchvision==0.22.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.1
transformers==4.55.0
typer==0.16.0
types-python-dateutil==2.8.19.14
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2023.3
tzlocal==5.2
ultralytics==8.3.174
ultralytics-thop==2.0.15
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.0.6
uvicorn==0.35.0
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.9
wrapt==1.16.0
yarl==1.20.1
zipp==3.17.0
zstandard==0.23.0
